Refactor: ImageDto, MenuDto 승격
